+++ src/system/mod.rs
@@ -3,3 +3,4 @@ pub mod coordinator_actor;
 pub mod landing_detection_actor;
 pub mod messages;
 pub mod simconnect_actor;
+pub mod simconnect_objects;